python_test/requirements.txt
Django==2.1.8
django-bootstrap4==1.0.1
django-filter==2.2.0
django-tables2==2.1.1